requirements/dev.txt
-f https://download.pytorch.org/whl/torch_stable.html                                                                                                                                                             
torch==2.2.2+cpu
torchvision==0.17.2+cpu
ultralytics==8.1.38
supervision==0.19.0
easyocr==1.7.1
yargy==0.16.0
pyspellchecker==0.8.1
grpcio==1.62.1
grpcio-tools==1.62.1
aiomonitor==0.7.0
pympler==1.0.1
ipykernel==6.29.4
ipython==8.23.0
PySide6==6.6.3.1
scikit-learn==1.4.1.post1
natsort==8.4.0
nltk==3.8.1
loguru==0.7.2
nats-py==2.7.2